[manage data] Add Curator intro (#487)
Closes https://github.com/elastic/docs-projects/issues/378

This PR adds content to the Curator page. I wish I could provide a
sentence or two on when it might be appropriate to use Curator instead
of or in addition to ILM, but it's not clear to me in what situations a
user might use Curator.

Note: I also removed the mapped pages from the frontmatter because all
redirects should go to the reference pages instead. I confirmed these
URLs are mapped in the reference content.

---------

Co-authored-by: wajihaparvez <wajiha.parvez@elastic.co>