[D&M] Add nuance to current comparison points between deployment types (#577)
- This is not a great page ATM and my changes don't make it good either,
they just fix a few things
- Table is ugly right now but it does make serverless sound less
limited.
- A couple of things were outdated or not exactly true
- Despite the branch name, I actually replaced the "Yes"iz and left Some
"No"z because in some cases the difference is rather "you can do it, but
it's manual" vs. "it's available but you need to enable it" vs. "it's
automatic or we're doing it for you trustme"

I think some capabilities/aspects are missing even though this does a
good job of showing that each option is mostly different in the degree
of customization/control the user has over what's under the hood

I'm ok with leaving this PR open until we figure out more things or how
we actually want to properly showcase these differences.

---------

Co-authored-by: shainaraskas <58563081+shainaraskas@users.noreply.github.com>