nodes orchestration updated for upgrade predicates (#697)
Moving the ECK upgrade predicates to reference documentation per
https://github.com/elastic/cloud-on-k8s/pull/8496#issuecomment-2704010367

- [Advanced control during rolling
upgrades](https://docs-v3-preview.elastic.dev/elastic/docs-content/pull/697/deploy-manage/deploy/cloud-on-k8s/nodes-orchestration#k8s-advanced-upgrade-control)

@shainaraskas : Let me know your thoughts, as this isn't exactly what we
have talked about :) I'd like to know if you like it this way or if you
prefer to keep the example and all narrative in this doc.

I think it's simpler this way, having only an introduction to the topic
and a link to the reference documentation for the list of predicates and
the use case example.

---------

Co-authored-by: shainaraskas <58563081+shainaraskas@users.noreply.github.com>