Removes redundant section  (#794)
As part of
[security-docs/6394](https://github.com/elastic/security-docs/issues/6394),
removes the malware and ransomware prevention sections from the
detection engine overview.

The initial request was to move these sections to the Endpoint
Protection section on the basis that they don't really fit in the
detection engine overview section, but based on my review of the
content, it seems to be already be [covered more comprehensively in the
Endpoint Protection
section](https://docs-v3-preview.elastic.dev/elastic/docs-content/pull/794/solutions/security/configure-elastic-defend/configure-an-integration-policy-for-elastic-defend#malware-protection),
and I think the page flows better with the content removed.

Preview: [Detection engine
overview](https://docs-v3-preview.elastic.dev/elastic/docs-content/pull/794/solutions/security/detect-and-alert)
(content was removed from here)