Fix references to Java client docs (#1071)
In https://github.com/elastic/elasticsearch-java/pull/988 we restructure
the 9.x Java client docs to re-instate the folder structure we had with
asciidoc along with adding docs for new 9.0 features.

Because of this, some references to the Java docs will need to be
updated once that PR is merged. This PR fixes this.

Co-authored-by: Marci W <333176+marciw@users.noreply.github.com>