Clarify ECH node bootlooping documentation (#1093)
## Description

Per an internal discussion (`sdh-cp #9409`), we would like to update the
ECH doc to clarify the node bootlooping related doc.

### Background

A setting is configured in Elasticsearch but it reports error

```bash
org.elasticsearch.common.settings.SettingsException: 
The configuration setting [xpack.security.authc.realms.foobar.foobar1.foobar2.client_secret] is required
```

And during the troubleshooting, we found the setting is configured via
client tooling and it's not properly configured (zero byte).

Based on that, we logically thought to solve the issue, users should 
- understand this may happen if the keystore setting
(`[xpack.security.authc.realms.foobar.foobar1.foobar2.client_secret]` to
be configured on ECH side) is invalid, missing or improperly configured
- understand the way to solve is to check if the keystore is valid or
not and make sure it's configured properly
- additionally if users configured this via client tooling, they should
try configuring this in cloud UI directly as a way to isolate the cause

### After PR being merged

The following orange section will be added


![image](https://github.com/user-attachments/assets/11210aa0-00d0-45ad-8e0c-2d6a15fb6b60)


![image](https://github.com/user-attachments/assets/bfc67644-a79e-465e-b4af-7a85d4650548)