Misc formatting fixes (#1115)
I did a broad search of all the generated HTML for a few patterns that
suggested formatting errors. Here are the fixes to the issues I found in
this repo.

Note: Some of the changes are subjective (for example, when to use
`subs=true` vs. just hardcoding a term like `Elasticsearch` in a code
comment).