Clarify user authentication methods not available for ECH (#1275)
### Description
Make it clear that LDAP, Active Directory and PKI is not configurable
for Elastic Cloud hosted environment.

### Background: 
- In an internal ticket (SDH - CP - 9505), we were guided by @beiske
that ECH doesn't support LDAP, but the public doc is not evident enough
for us to understand this fact
- We agreed that a doc PR to clarify this would be necessary

I also found Active Directory and PKI are the same that are not
applicable on ECH.
So I made a quick doc PR to make these related user authentication docs
clear.


### After PR is merged

The orange sections will be added


![image](https://github.com/user-attachments/assets/37637d8c-366f-4109-9359-219abe0eb0c0)

![image](https://github.com/user-attachments/assets/aa64b9f3-3ad0-42d3-86a2-773ef09cf722)

![image](https://github.com/user-attachments/assets/d2df7c04-0e54-4b68-ba0e-552282ec9bbe)

![image](https://github.com/user-attachments/assets/c3400d4b-33df-4168-befc-6548961cf0e0)


### Additional Notes

To doc team: Maybe we could also use `note` instead of `warning`. I
don't have strong opinion here. Would leave this to your capable hands
for decision 🙏

---------

Co-authored-by: Florent Le Borgne <florent.leborgne@elastic.co>