Reworks 'Generating alerts for anomaly detection jobs' guide' (#1305)
### Context

As part of the rework of the [ML anomaly detection
book](https://www.elastic.co/guide/en/machine-learning/current/ml-ad-overview.html)
a new
[How-tos](https://www.elastic.co/guide/en/machine-learning/current/anomaly-how-tos.html)
section was introduced. Some of the content currently listed there does
not fully follow a how-to structure and needs to be updated to better
align with the intended format.

### About this PR

This is the first PR contributing to the
https://github.com/elastic/search-docs-team/issues/27 issue for
improving the How-to guides.

In this PR, I've restructured the **Generating alerts for anomaly
detection jobs** guide to make it more task-focused and action-oriented.
Updates include:

- A clearer introduction
- Simplified and structured text
- Defined goals and prerequisites
- Step-by-step instructions in an instructional style
- Updated screenshots for 9.0 changes

As part of this effort, I'm also working on creating a shared How-to
guide template to support consistency across all guides:
https://docs.google.com/document/d/1dfbmAf4gZsNbY33gFKATRsIxJNJHqgN9C_BYZ2Jhff8/edit?usp=sharing

### Preview

- [Generating alerts for anomaly detection
jobs](https://docs-v3-preview.elastic.dev/elastic/docs-content/pull/1305/explore-analyze/machine-learning/anomaly-detection/ml-configuring-alerts)