Debian pkg config file updated in snippet (#1457)
Minor change to fix the package configuration file in the Debian
installation guide, which was incorrectly set to the RPM based file.

@shainaraskas : As the content came from a reused snippet we have lost
the code formatting. Let me know your preference here. If you want to
not use the `snippet` for this block of content or if we keep using it
and we lose the code formatting.

If we use \`{{pkg-conf}}\` the substitution doesn't work.

I'm going to write a docs-builder issue to see if we can support inline
code blocks with substitutions.

Closes https://github.com/elastic/docs-content/issues/1430

---------

Co-authored-by: shainaraskas <58563081+shainaraskas@users.noreply.github.com>