yaml values indentation fixed and aligned with the other oidc page (#1493)
This solves a bug discovered @rgshenoy in
https://github.com/elastic/docs-content/pull/1416
Replaces https://github.com/elastic/docs-content/pull/1416

Some indentation in the yaml configuration for Elasticsearch and Kibana
was incorrect. We have fixed it and aligned the format with the other
OIDC document.

Note that while reviewing all other examples in other realms (LDAP,
SAML, etc) there's no alignment at all in the format of the suggested
yaml settings. It might be good to align them in a separate PR, although
it's not really an issue or important.

cc: @shainaraskas