add remote cluster traffic filter info in remote-cluster connection page (#1478)
This pull request updates the documentation for Elastic Cloud remote
clusters to include a note about traffic filtering requirements when it
is enabled. The note provides guidance for administrators on configuring
a traffic filter for remote clusters.

Documentation updates for traffic filtering:

*
[`deploy-manage/remote-clusters/ec-remote-cluster-other-ess.md`](diffhunk://#diff-7e34b42994e11a4cb6ebbff3e91b94d3bce843ff14b18fd7876e6d821d10d148R239-R242):
Added a note explaining that when traffic filtering is enabled, the
remote cluster administrator must configure a traffic filter of type
remote cluster using the organization ID or Elasticsearch cluster ID. A
link to detailed instructions is also provided.
*
[`deploy-manage/remote-clusters/ec-remote-cluster-same-ess.md`](diffhunk://#diff-7981e127338b90c990011ed12b1da648a4eab1c554c97f5c5dcb8c12894b6110R276-R279):
Added the same note about traffic filtering configuration as in the
previous file, ensuring consistent guidance across documentation for
remote clusters.
https://elasticco.atlassian.net/browse/ECPTRAFFIC-546

---------

Co-authored-by: Edu González de la Herrán <25320357+eedugon@users.noreply.github.com>