clarify MFA for team use case (#1576)
## Description

Clarify MFA for team use case by making it clear that the only secure
and recommended way is to add members to the organization.

## Background & Motivation

Background is per
https://github.com/elastic/sdh-control-plane/issues/9470#issuecomment-2820959159
and https://github.com/elastic/support-tech-lead/issues/1564, which I
had with @brunofarache internally that we should clarify the docs to
make it clear that the only recommended way is to add members to the
org.

## Before / After PR merged

:: Before 

https://www.elastic.co/docs/cloud-account/multifactor-authentication

<img width="1501" alt="image"
src="https://github.com/user-attachments/assets/60b6ea51-d18f-4837-9031-803e7ec4e822"
/>


:: After

Orange part will show up.


![image](https://github.com/user-attachments/assets/d18617f5-a737-44f4-a229-591067b57280)

---------

Co-authored-by: shainaraskas <58563081+shainaraskas@users.noreply.github.com>