Update prepare-to-upgrade.md (#1604)
Adding a small recommendation that we have always shared for major
upgrades.

If the user is already on the latest minor (8.18.x) it's also
recommended to upgrade to 8.18.latest before upgrading to 9.x, as there
could be fixes and improvements on upgrade assistant and other
components.

A clear example of this could be
https://github.com/elastic/stack-docs/pull/3024

---------

Co-authored-by: shainaraskas <58563081+shainaraskas@users.noreply.github.com>