Small ES|QL editor changes for 9.1 (#1818)
Batching a few changes for the various ES|QL editor improvements that
don't require a lot of docs.
The main need here for newly supported commands and functions is to make
the ES|QL reference quicker to find for anyone reading these docs.

- Adding a direct link to the ES|QL syntax docs from a few pages
- Adding a section about keyboard shortcuts supported by the editor
- Moving the LOOKUP_JOIN section from Discover docs to ES|QL editor docs
as it's not Discover specific, and to prepare for future improvements
around this command (file upload...)

closes: https://github.com/elastic/docs-content/issues/1677
closes: https://github.com/elastic/docs-content/issues/1678
closes: https://github.com/elastic/docs-content/issues/1680
closes: https://github.com/elastic/docs-content/issues/1681

Other ES|QL-related docs content changes are in the works and will be
treated separately:
- https://github.com/elastic/docs-content/issues/1810
- https://github.com/elastic/docs-content/issues/1679
- https://github.com/elastic/docs-content/issues/1669