ECE redirect fix - Elasticsearch settings (#2009)
Adds
https://www.elastic.co/guide/en/cloud-enterprise/current/ece-change-user-settings-examples.html
to the appropriate doc as a mapped_page.

This is part of the work for
https://github.com/elastic/docs-projects/issues/494

In a parallel activity we will request the redirect of that page to
https://www.elastic.co/docs/deploy-manage/deploy/cloud-enterprise/edit-stack-settings-elasticsearch

In this case we haven't completely imported / migrated the original
document because we don't want to duplicate a random list of settings
which are available in the reference doc and also in the narrative docs
about specific features.