ECPTRAFFIC-1376: Update traffic filter docs to remove traffic filter inconsistency (#2063)
Updated the documentation for using traffic filters with remote
connections that stated that traffic filters cannot be used with API key
auth. This is incorrect and left users with no option given that TLS
Certificates are deprecated and traffic filters are not enforced for
that auth method anyway. I also added clarification for the region TLS
cert callout in the event there is confusion with that only being
relevant for TLS Certificate auth (deprecated).

Ticket: https://elasticco.atlassian.net/browse/ECPTRAFFIC-1376