MinIO updates in ECE docs (#1996)
In relation with https://github.com/elastic/cloud/issues/134936, we have
realized that the doc
`/deploy-manage/tools/snapshot-and-restore/minio-on-premise-repository.md`
was really outdated and needed some love.

This is the main page to review:

[MinIO self-managed
repository](https://docs-v3-preview.elastic.dev/elastic/docs-content/pull/1996/deploy-manage/tools/snapshot-and-restore/minio-on-premise-repository)

Applied changes:
- Changed wording to `MinIO` across all docs. No minio or Minio anymore.
- Fixed a lot of broken links to MinIO docs.
- Removed screenshots (thanks @Kushmaro for pointing that out), and
replacing them by proper links and instructions.
- Adding a note to recommend installing MinIO in separate hosts,
although we support installing it in ECE allocator hosts.
- Removed the manual steps to restore a snapshot from another cluster as
now that's offered and documented as an ECE feature. I've linked to the
right doc for that.
- Improved the structure of certain sections as they were hard to
understand (the bucket creation step for example was incorrect).
- Renamed `offline` to `air-gapped` (for consistency with the new IA and
new docs).
- Changed title from `on-premise` to `self-managed` (because ECE hosts
or the customer minio hosts could be on-premises or on cloud VMs
technically. What they are in reality is `self-managed`).
- Added useful links to other parts of the documentation.
- Removed the statements about 6.x and 5.x versions of Elasticsearch.
These are unneeded as in ECE 4.0 it's mandatory for all deployments to
be on 8.0.0 or above. So there's no need to mention it.

Closes https://github.com/elastic/cloud/issues/134936

---------

Co-authored-by: David Kilfoyle <41695641+kilfoyle@users.noreply.github.com>