snapshots clients background added for Azure and S3 (#2148)
This PR adds a paragraph to clarify the meaning of `clients` in the
context of the documents for Azure and S3 repositories.

We have received
[feedback](https://github.com/elastic/docs-feedback/issues/45) for the
Azure repository document showing frustration as it wasn't clear what
the azure client is and if it's something external to Elastic that the
customer needs to take care of.

Closes https://github.com/elastic/docs-feedback/issues/45

---------

Co-authored-by: David Kilfoyle <41695641+kilfoyle@users.noreply.github.com>