Replace placeholder URLs (#2141)
In light of finding abuse of dummy links in our docs, this replaces
placeholder URLs with non-hostable segments.

A few thoughts:
- I noticed that for Cloud API examples we sometimes use
`COORDINATOR_HOST` rather than the variable `$COORDINATOR_HOST`. The
latter is safer, and also we explicitly tell users to export that
variable in their shell to run the curl API examples.
- Some other changes are just for consistency, i.e., using angle
brackets for segments that need to be replaced.
- Where possible I removed `http` and `https` from dummy URLs, but in
some cases it's explicitly the protocol that's being demonstrated, so I
left those and changed the url content to contain angle brackets.

Rel: https://github.com/elastic/docs-content/issues/1799

---------

Co-authored-by: shainaraskas <58563081+shainaraskas@users.noreply.github.com>