ece medium and large installation instructions improved (#2146)
This PR fixes and improves ECE medium and large installation
instructions.

Fixes:
- The description of amount of allocators needed for a large
installation wasn't coherent.
- We had a [bug](https://github.com/elastic/cloud/issues/133321) in the
procedure where we were not adding the step of removing the allocator
role from the initial host, which is important.
- Large instructions also needed to remove proxy role.
- Removing proxy role on first host has the implication of needing to
update the endpoint URLs.

Enhancements:
- Added a small note to understand why the drawing shows the allocator
role `greyed out` in one of the hosts. It improves the reading flow.
- Log into Cloud UI was in the wrong order :)
- Change deployment configuration link points now to the right place.

Closes https://github.com/elastic/cloud/issues/133321

---------

Co-authored-by: David Kilfoyle <41695641+kilfoyle@users.noreply.github.com>