Docs for high contrast mode in Kibana (#1986)
Update0 - 10/07 Currently checking with eng whether there are deployment
type restrictions that apply to this because I could not test it
successfully on cloud and serverless
Update 1 - 15/07: Holding as the team is trying to get this to work also
on cloud deployments
Update 2 - 21/07: The option was also added to the Appearance menu
available in cloud instead of the user profile's menu, so it's now
correctly available in Kibana no matter the deployment mode used.

I've put the new page next to the dark mode one under "Cloud account and
preferences" as it sounds like the current best place for personal
settings.

Closes: https://github.com/elastic/docs-content/issues/1675