[Private preview] Add docs for Cloud Connect + AutoOps for self-managed clusters (#2287)
The docs added in this PR are meant for **private preview**, so they are
hidden from the ToC.

I've added all the docs at the same level for now, because hidden files
[can't have child
pages](https://docs-v3-preview.elastic.dev/elastic/docs-builder/tree/main/configure/content-set/navigation#hidden-files).
We can fix that and decide on exact positioning for GA. It doesn't
matter right now because customers won't be able to _navigate_ to them
from the ToC anyway.

To avoid any confusion about the structure, I've added a section
overview on the **AutoOps for self-managed clusters** page with links to
all the other pages.

Closes: https://github.com/elastic/docs-content-internal/issues/16 and
https://github.com/elastic/docs-content-internal/issues/12