edit defend advanced options wording (#2157)
This edits the Defend advanced option wording with the following primary
goals:

1. Standardizing some word choices and stylistic choices across all the
options.
2. Shortens the _italics_ text. This is the text that also appears in
Kibana tooltips. Once this PR is approved I'll open a Kibana PR to align
those tool tips to the text agreed to here.

Originally we'd
[agreed](https://github.com/elastic/docs-content/pull/1445#issuecomment-2949329797)
to completely remove the _italics_ portion on this page and make sure
the Kibana tool tips were "short" with an expectation users turn to this
more detailed documentation when needed. I tried, but was only as
successful as what I'm proposing to merge in this PR (usually shorten
the _italics_ tool tip and always still show it on elastic.co).

Thoughts @natasha-moore-elastic @gabriellandau @roxana-gheorghe @nfritts
@joe-desimone

# [Docs Preview as of July
28](https://docs-v3-preview.elastic.dev/elastic/docs-content/pull/2157/reference/security/defend-advanced-settings)

---------

Co-authored-by: natasha-moore-elastic <137783811+natasha-moore-elastic@users.noreply.github.com>