Add "Contact support" to EDOT Troubleshooting (#2256)
This PR adds a new troubleshooting page that guides users on how to
contact Elastic Support when troubleshooting issues with the EDOT
Collector or SDKs.

The page includes a checklist of information users should gather before
opening a case. This addition is meant to reduce support back-and-forth
and help users submit actionable support requests.