Update tail-based-sampling.md (#1360)
A minor change to the data type for `sample_rate`. We mention that

> The sample rate must be greater than or equal to `0` and less than or
equal to `1`.

which if restricted to an `int` type would only allow `0` and `1`. A
more reasonable (unless there is a reason not clear to me) data type
here would be `float`, allowing any decimal precision value, like the
`0.01` mentioned as example.

Co-authored-by: Benjamin Ironside Goldstein <91905639+benironside@users.noreply.github.com>