Security AI Assistant access to Elastic Docs in air-gapped envs (#2598)
Fixes #2404 by documenting the process for enabling AI Assistant to
access Elastic documentation from an air-gapped environment. The process
is a bit different from that for the Obs/Search Assistant, since the
Assistants' settings UIs are different, so I confirmed the process for
the Security AI Assistant directly with someone who works on it.

---------

Co-authored-by: natasha-moore-elastic <137783811+natasha-moore-elastic@users.noreply.github.com>