[Response Ops][9.x & Serverless]: Improve docs for recovered status (#2597)
## Summary

Contributes to https://github.com/elastic/docs-content/issues/2313 and
https://github.com/elastic/docs-content/issues/2143 by elaborating on
the logic behind the recovered alert status and providing examples of
when statuses change from active -> recovered and flapping -> recovered.
Rule docs for Observability and Stack Management are updated.

This PR also makes the following minor changes:
- Observability page updates:
- Added `applies to` tags to the front matter to show that the content
is relevant to Stack and Serverless users.
- Moved the alert status docs to their own dedicated section for easier
scanning and reference. This also makes the docs for alert statuses
linkable if we ever need to reference the docs from the UI.
- Minor revisions to the description for the flapping status. Also
updated the note to show that alert flapping is enabled by default can
can be configured if users want to modify the conditions for the
flapping status.
- Replaced old `svg` images of icons with the formal ones
[here](https://elastic.github.io/docs-builder/syntax/icons/).

**Corresponding 8.x doc updates**:
- Observability -
https://github.com/elastic/observability-docs/pull/4949
- Kibana - https://github.com/elastic/kibana/pull/232301

## Previews
[Available
here](https://github.com/elastic/docs-content/pull/2597#issuecomment-3192227447)