Ingest pipeline best practices (#1381)
based on the discussions here:
https://github.com/elastic/docs-content/issues/1052

this is my first PR against the docs, and I am building a couple of new
pages. I think it makes sense to split it out. I am putting it into that
part of the docs.
https://www.elastic.co/docs/manage-data/ingest/transform-enrich/ingest-pipelines
The tips and tricks are generic and not specific to just o11y, or
security.

<img width="1776" alt="image"
src="https://github.com/user-attachments/assets/f5254116-7170-42b1-911d-1387605fbb8b"
/>

---------

Co-authored-by: Colleen McGinnis <colleen.mcginnis@elastic.co>
Co-authored-by: Colleen McGinnis <colleen.j.mcginnis@gmail.com>