Corrects use of autorefresh option (#2712)
The author of #1745 is correct, `autorefresh = 1` is not a config option
that's available in dnf or yum. However, it is a valid and often used
option in
[zypper](https://doc.opensuse.org/documentation/tumbleweed/zypper/)
(openSUSE distros).

I've added tabs to account for this discrepancy (when creating the repo
definition) and using tabs actually fixes the next step as well, which
is trying to account for multiple Linux distros commands to install
Elasticsearch from an RPM repo.
The RHEL-based commands for installing will also differ based on the
version or RHEL you're using, so that tab accounts for RHEL 8 + (when
dnf was introduced) and RHEL 7 & earlier for yum.

I've also updated the names of the distros for accuracy:
* Red Hat Enterprise Linux (RHEL)
* openSUSE
* SUSE Linux Enterprise Server (SLES)
* Oracle Linux (was renamed from Oracle Enterprise)
* CentOS

Fixes #1745