Replaces all links to point to the deduplicated/reconciled roles page (#2794)
The Elasticsearch [Built-in roles page in the Deploy and manage
section](https://www.elastic.co/docs/deploy-manage/users-roles/cluster-or-deployment-auth/built-in-roles)
is being replaced with its [counterpart in the Reference
section](https://www.elastic.co/docs/reference/elasticsearch/roles).
This PR replaces all links from within the narrative docs to the correct
(deduplicated and reconciled) page.

Relates to #2738 and
[#133831](https://github.com/elastic/elasticsearch/pull/133831)