Updates MITRE docs to currently used version `v17.1` (#2518)
🟢 **NOTE TO SELF:** Merge this on Tuesday, Sep 3, 20205 and then open
PRs for https://github.com/elastic/docs-content/issues/2791

## Summary

Updates the 9.x and Serverless docs to show that detection rules will
use the MITRE ATT&CK® version v17.1 in 9.2 and next weeks' Serverless
release.

Because we need to show that earlier versions of 9.x (specifically
9.0.0-9.0.6 and 9.1.0-9.1.3) use an older version of MITRE ATT&CK®
(v16.1), I created a table to show how the versions are mapped.

## Related
- Doc issue: https://github.com/elastic/kibana/issues/166152
- Dev PR: https://github.com/elastic/kibana/pull/231375

## Preview


https://docs-v3-preview.elastic.dev/elastic/docs-content/pull/2518/solutions/security/detect-and-alert/mitre-attandckr-coverage

---------

Co-authored-by: Nastasha Solomon <nastasha.solomon@elastic.co>
Co-authored-by: Nastasha Solomon <79124755+nastasha-solomon@users.noreply.github.com>