[ResponseOps][Docs]: Maintenance window GAing and category options removed (#2855)
Closes https://github.com/elastic/docs-content/issues/2796. Updates the
maintenance windows docs as follows:
- Updates the page-level `applies to` tags to show that the feature is
in tech preview for 9.0-91 and GA in 9.2 and Serverless.
- Removed the note at the top of the page that showed the feature was in
tech preview.
- Adds a note that explains that category-specific maintenance window
options are unavailable (removed) in 9.2 and Serverless. Also updated
the first item in the note further down that explains how you can select
only a single category when you turn on filters.
- Refreshes both images on the page since they show elements that are no
longer available in 9.2 and Serverless (the tech preview label and the
option to choose categories for the maintenance windows

Preview:
https://docs-v3-preview.elastic.dev/elastic/docs-content/pull/2855/explore-analyze/alerts-cases/alerts/maintenance-windows