Adds an xlink to the toc that points to the reconciled privileges page (#2989)
As part of replacing the Elasticsearch privileges page in the [Deploy
and manage
section](https://www.elastic.co/docs/deploy-manage/users-roles/cluster-or-deployment-auth/elasticsearch-privileges)
with the equivalent page in the [Reference
section](https://www.elastic.co/docs/reference/elasticsearch/security-privileges).

This PR will retire the existing page in the **Deploy and manage**
section and instead link to the correct, deduplicated page in the
**Reference** section.

I've also updated some links that were referencing this page.

Relates to #2218 and #2738

-------

Here's the
[preview](https://docs-v3-preview.elastic.dev/elastic/docs-content/pull/2989/deploy-manage/users-roles/cluster-or-deployment-auth/user-roles)
for the xref-ed page (Click **Elasticsearch privileges** in the nav):

<img width="1149" height="559" alt="image"
src="https://github.com/user-attachments/assets/a5eaf146-0aaa-4ea2-bc74-6122be56986d"
/>