Updates to 'Migrate to ILM on self/ECK' (#2999)
Part of [Edu's review of the ILM
docs](https://github.com/elastic/docs-content-internal/issues/190?reload=1?reload=1).

This updates the recently renamed "[Migrate to ILM on self-managed and
Elastic Cloud on Kubernetes
deployments](https://docs-v3-preview.elastic.dev/elastic/docs-content/pull/2999/manage-data/lifecycle/index-lifecycle-management/manage-existing-indices)"
page, to:
 - Change the "Curator" mention to a link
 - Clarify that the scope of the page doesn't include data streams
 - Add a link to some quite helpful rollover examples

As part of that same discussion it was suggested that we update
"[Manually apply a lifecycle policy to an
index](https://docs-v3-preview.elastic.dev/elastic/docs-content/pull/2999/manage-data/lifecycle/index-lifecycle-management/policy-apply)"
with a link to the "Migrate to ILM" section, which is a similar but
slightly more elaborate use case. So I've included that as well.