Removes agentless integration deployment beta list (#2998)
Fixes #2995 by removing the list of beta agentless integrations and
making some related updates. This list had become outdated, and
agentless deployment for integrations will be GA in 9.2 anyway. This
page will be further updated when 9.2 is released, including
highlighting the ability to filter for integrations that support
agentless deployment, and clarifying that agentless deployment is GA.

---------

Co-authored-by: Fabrizio Ferri-Benedetti <fabri.ferribenedetti@elastic.co>
Co-authored-by: florent-leborgne <florent.leborgne@elastic.co>