ECK: Add new Fleet recipe (#2699)
Depends on https://github.com/elastic/cloud-on-k8s/pull/8788

Add a new recipe for ECK managed Fleet servers after a recent support
escalation that illustrated that users would benefit from additional
guidance on how to set this up.

> This example shows how to expose the Fleet Server to the outside world
using a Kubernetes Ingress resource. The Fleet Server is configured to
use custom TLS certificates, and all communications are secured with
TLS. The same Fleet Server is also accessible from within the cluster,
allowing agents to connect to it regardless of their location. Refer to
the comments in the `fleet-ingress-setup.yaml` file for more details on
how to set up the Ingress resource and TLS certificates to enable this
configuration.