Adds to agentless FAQ about upgrading agentless integrations (#3108)
Fixes #3029 by adding a new FAQ item about upgrading agentless
integrations to new versions. There's a fix for 9.2, but prior to 9.2
the UI wouldn't allow you to upgrade an existing integration, other than
by deleting and reinstalling it. I decided not to mention serverless in
the FAQ item because the upgrade issue has been fixed in serverless.