[Security] [Serverless: Sep 23] Endpoint Exceptions feature privilege in Stack (#2983)
Resolves https://github.com/elastic/docs-content/issues/2974.
Documents the addition of the **Endpoint Exceptions** sub-feature
privilege in Stack for 9.2. Until now, this privilege only existed in
serverless.

For consistency, this PR also moves the location of the {applies_to}
badge for the **Global Artifact Management** privilege, in line with the
[badge placement
guidelines](https://elastic.github.io/docs-builder/contribute/cumulative-docs/badge-placement/#table-row).

---------

Co-authored-by: Visha Angelova <91186315+vishaangelova@users.noreply.github.com>