Tweak agent builder's llm selection details (#3051)
[Emma
confirmed](https://elastic.slack.com/archives/C056H3WA1BL/p1758189381969239?thread_ts=1757949389.230719&cid=C056H3WA1BL)
that the setting to change the default LLM is _not_ going to be
available before 9.2.0, so we shouldn't reference it in the first docs.

We shouldn't document the non-UI path, as the API to change this setting
is `/internal/`. We'll be OK to communicate this directly to the
customers we invite to the private preview, but otherwise we shouldn't
document this until the setting is available to use in the UI.

Also fixed a typo.

Also took out an explicit reference to Claude 3.7, after confirming with
@joemcelroy and @maxjakob that we don't want to advertise exactly what
the Elastic LLM is using under the hood.