Add `kibana.alert.original_data_stream` to the list of alert schema fields (#3011)
@rylnd @yctercero let me know if I got this right or suggest anything
you'd like to see. The original issue mentions that
`kibana.alert.original_event.*` behaves similarly but it's not described
much in the docs from what I could find, so we may also clarify things
for that one if you feel it's needed.

Once we're good with the content here, I'll create the corresponding PR
for 8.19, as docs for this version are still in asciidoc in their
previous location.

Closes: https://github.com/elastic/docs-content/issues/2673