Update versioning and availability (#3004)
Relates to https://github.com/elastic/docs-content/pull/2795

This PR updates the introductory [Versioning and
availability](https://www.elastic.co/docs/get-started/versioning-availability)
page.
It is a work in progress until we address the following feedback:

-
https://github.com/elastic/docs-content/pull/2795#discussion_r2321576573:
"I don't know if this [Examples of where availability can vary] section
is very clear. I think it could use some more details or introductory
content. Somewhere on this page, we should try to explain a few things,
how these multiple, differently versioned, products work together and
how this can impact availability:
- Elastic Cloud, ECE, and ECK are orchestrators used to create and
manage serverless projects (Elastic Cloud only) and deployments of the
Elastic Stack
- The versions of the Elastic Stack that you can deploy depend on the
version of these orchestrators. For example, Elastic Stack 9.0 is not
available to deploy on ECE 3.1.
- Because some products are released on different schedules and have
different architecture or network constraints, it is possible that a
feature isn't available for some deployment types, or becomes available
at a different time for each deployment type."
-
https://github.com/elastic/docs-content/pull/2795#discussion_r2322287185
"we also do some of this explanation on the following surfaces and you
can mine them for language / make sure they align:
- https://www.elastic.co/docs/deploy-manage/deploy#about-orchestration
-
https://www.elastic.co/docs/deploy-manage/deploy#versioning-and-compatibility
    - https://www.elastic.co/docs/deploy-manage/upgrade/plan-upgrade"
-
https://github.com/elastic/docs-content/pull/2795#discussion_r2322287185
"> Because some products are released on different schedules... would be
careful with this because a lot of functionality is controlled by the
stack version. might indicate that these are usually at the admin layer
rather than the storage/querying layer of elastic. examples would be
helpful (e.g. not all auth types are available for all deployment types
[e.g.
[ldap](https://www.elastic.co/docs/deploy-manage/users-roles/cluster-or-deployment-auth/ldap)]
- that's kind of a layer "above" / "around" the es engine)"

---------

Co-authored-by: Janeen Roberts <janeen.roberts@elastic.co>
Co-authored-by: Janeen Mikell Roberts <57149392+jmikell821@users.noreply.github.com>
Co-authored-by: Fabrizio Ferri-Benedetti <fabri.ferribenedetti@elastic.co>